sense7ai/requirements.txt
django==2.2.7
djangorestframework